I restructured my code due to I was not passing the test. the previous code definitely works it just does not pass all the tests. i redid so that assignment can pass all test
