+++ app.js
@@ -1,9 +1,11 @@
 const express = require("express");
 const app = express();
+const cors = require("cors");
 
- const logsController = require("./controllers/logsController.js");
+const logsController = require("./controllers/logsController.js");
 
 app.use(express.json());
+app.use(cors());
 
 app.use((request, response, next) => {
   next();
@@ -13,7 +15,7 @@ app.get("/", (request, response) => {
   response.send("Welcome to the Captain's Log!");
 });
 
- app.use("/logs", logsController);
+app.use("/logs", logsController);
 
 app.get("*", (request, response) => {
   response.status(404).json({ error: "page not found" });
