+++ package.json
@@ -6,8 +6,7 @@
   "scripts": {
     "test": "jest",
     "start": "node server.js",
-    "dev": "nodemon server.js" 
-    
+    "dev": "nodemon server.js"
   },
   "keywords": [],
   "author": "",
@@ -18,6 +17,7 @@
     "supertest": "^6.1.3"
   },
   "dependencies": {
+    "cors": "^2.8.5",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "test": "^3.3.0"
